requirements.txt
requests
sqlalchemy
transmission_rpc
python-dotenv
urllib3>=1.26.5
inquirer
argparse
fastapi
uvicorn
databases